Add: :sparkles: browserync + Concurrently working with Parcel
Using Browsersync & parcel on the package.json for npm run dev is not
working. This is due to we have to tell explicit which of them package
have to run first in order to get it to work. By trying to work together
is not going anywhere.

Solution is to use a dependency which is called Concurrently:
https://www.npmjs.com/package/concurrently

This allow us to run multiple commands concurrently like npm run
watch-js & npm run watch-less but better.

We can actually run "parallel" both of them. This will ensure
Browsersync and Parcel can do their own seperate job together.

- Browsersync will ensure to hot reload every PHP files we change.
- Parcel will do Hot Module Reload (HMR) for SCSS + JS.

We get best of the both world!