Merge: 🔀 'feature/_gridlex.scss' to main
Add: ✨ Gridlex Vendor

This Grid dependency is very intuitive and easy to use.
This has the 3rd most highest star of Git bookmark purely dedicated for
CSS grid. It has tons of options to make your website responsive.
It's "Simple. Robust. Responsive. Flexible. Built with Sass."

The concept is simple:
you need to wrap your .col in a .grid.

That's all.

Source:
https://github.com/search?o=desc&q=scss+grid&s=stars&type=Repositories
https://github.com/devlint/gridlex
http://gridlex.devlint.fr/