Hotfix: 🚑 Parcel Invalid Version: undefined
The reason we are getting this issue is due to the
Babel API has been changed, and is breaking the Parcel down.

Parcel v1 is not maintained anymore, and the V2 is the new Parcel.
We have to find a way to work around this case to make sure
we don't get the error of Invalid Version: undefined.

In this commit. I have added

{
  "name": "my-package",
  "scripts": {
    "preinstall": "npx npm-force-resolutions"
  },
  "resolutions": {
    "@babel/preset-env": "7.13.8"
  }
}

This will make sure that we are forcing our package-lock.json to always
stay intact with the @babel/preset-env version 7.13.8.

Next time in order to start the WP Starter Theme is to:

1. npm install, but do it twice
2. or do yarn install

After that, run npm run dev or npm run build