Add: ✨ Compress-images dependency
Images files needs to be added on src/img/ folder in order for it to work.
When compressing, it will create a temporarily folder which is called
temp_folder. All of these image files that has been compressed will
be moved to the temp_folder. When it's finished, then it will take
all the image files of temp_folder and move back to the original
src/img/ folder.

Compressing engines we are using for the images are:

png = pngquant
jpg = mozjpeg
gif = gifsicle
svg = svgo

Source:
https://www.npmjs.com/package/compress-images