Add: ✨ Github Actions for deployment staging and production
This file is used to compile our npm run prod and deploy to a staging
or production server.

Currently the setup of the steps won't work of course. This needs
be configured based on the project name. What needs to be configured are:
- username@remote_host
- _theme_name
- _site_name

The _site_name is the folder name when connecting to
ssh username@remote_host.

The PRIVATE_SSH and KNOWN_HOST needs to have the values in secret of
your project name of GitHub Repository.