Add: ✨ Using RichText Component in the Editor for outputting front-end
We could output the return as <p>{content}</p> to show it on the front-end,
but the issue is what if the user actually used the Bold text?
The <p> tag element wouldn't work.

In order to get it to work, we have to use <RichText.content value={content}/>
in order to work