Add: ✨ Grid Component for Team Memmbers. Acting like an inner block
The Grid Component is supposed to be a block where TeamMembers Component
will live in. The <InnerBlock /> is from WP Gutenberg block pre-made.

This InnerBlock will make sure you can add inner stuff of your component.

The 'allowedBlocks' attribute will ensure which blocks you can be used
specifically only. Think of it like a "filter" of which blocks is
allowed to be added inside