Add: ✨ Handling Blocks with Unfinished Image Uploads
When uploading a large image, it will take time to finish it.
If for some reason the user updating or publishing the post way to fast
before the image is finished to upload, and we try to refresh the editor again
The image will not be uploaded and there is a side effect that shows that
loading spinner is still loading because the image creates an temporarily
blob image url which is only valid on the current window of the editor
when the user tried to upload the image before publishing/updating.

This commits ensure that this bug will not show spinner if user for some
reason clicking on the update/publish way to fast. Instead, it will
just show the normal media upload